Impelement parts of new design.
I went back to the drawing board and now have a much better
understanding of how the canvas stuff works, as well as some ideas for
handling multiple "objects" being drawn onto it at once. I'm getting
some errors that I need to handle though. I also need to implement the
actual drawing of objects in an array, as well as picking a random word
to add to the screen.